Refactor: move project row to shared file
* Update import paths
