+++ apps/bestofjs-nextjs/src/app/api/og/route.tsx
@@ -2,13 +2,7 @@ import { api } from "@/server/api-remote-json";
 import { getHotProjectsRequest } from "@/app/backend-search-requests";
 
 import { ImageLayout } from "./og-image-layout";
-import {
-  Box,
-  ProjectLogo,
-  StarIcon,
-  generateImageResponse,
-  mutedColor,
-} from "./og-utils";
+import { Box, ProjectRow, generateImageResponse, mutedColor } from "./og-utils";
 
 export const runtime = "edge";
 
@@ -32,56 +26,6 @@ export async function GET() {
         ))}
       </Box>
     </ImageLayout>
-  );
-}
-
-export function ProjectRow({
-  project,
-  rank,
-}: {
-  project: BestOfJS.Project;
-  rank: number;
-}) {
-  return (
-    <Box
-      style={{
-        color: "white",
-        gap: 24,
-        alignItems: "center",
-        borderBottom: "1px",
-        borderColor: "#3d3d42",
-        borderStyle: "dashed",
-        borderTopWidth: rank === 1 ? 1 : 0,
-        padding: "8px 0",
-      }}
-    >
-      <Box style={{ color: mutedColor }}>#{rank}</Box>
-      <Box>
-        <ProjectLogo project={project} size={80} />
-      </Box>
-      <Box
-        style={{
-          justifyContent: "space-between",
-          flex: 1,
-        }}
-      >
-        <div>{project.name}</div>
-        <Box>
-          <ShowStars project={project} />
-        </Box>
-      </Box>
-    </Box>
-  );
-}
-
-function ShowStars({ project }: { project: BestOfJS.Project }) {
-  return (
-    <Box
-      style={{ flexDirection: "row", alignItems: "center", color: mutedColor }}
-    >
-      <Box>{`+${project.trends.daily}`}</Box>
-      <StarIcon />
-    </Box>
   );
 }
